+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 - Handle aliases in mail regex for b2b sale tunnel
+- Add next_url configuration for OpenEdX Hawthorn login/register redirects
 
 ### Changed
 
@@ -30,7 +31,6 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Handle French and English language in Storybook
 - Upgraded js dependencies, fixed tests and linting errors
-- Display batch order enrollment status in dashboard
 
 ### Fixed